Переделал библиотеку элементов цепи, добавив возможность добавления и расчета сложных цепей(состоящих из паралельных и последовательных подцепей и элементов)
