--- МЕХ/Circuit/CircuitCalculator.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2026
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CircuitCalculator", "CircuitCalculator\CircuitCalculator.csproj", "{1BB0B5AC-EB10-47CC-8597-1B23D7AE028C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CircuitElements", "CircuitElements\CircuitElements.csproj", "{F1FCBCA1-614D-43D3-B409-4461882EA899}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1BB0B5AC-EB10-47CC-8597-1B23D7AE028C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1BB0B5AC-EB10-47CC-8597-1B23D7AE028C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1BB0B5AC-EB10-47CC-8597-1B23D7AE028C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1BB0B5AC-EB10-47CC-8597-1B23D7AE028C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FBB8E4C8-64C3-4390-8F8E-49DCF371FBFF}
-	EndGlobalSection
-EndGlobal
