--- МЕХ/Circuit/CircuitCalculator/Controls/DrawingElements.cs
@@ -1,11 +1,0 @@
-﻿namespace CircuitCalculator.Controls
-{
-	/// <summary>
-	///     Перечисление элементов цепи, неучаствующих при расчетах
-	/// </summary>
-	public enum DrawingElements
-	{
-		StartingElement,
-		FiniteElement
-	}
-}
