--- МЕХ/Circuit/CircuitCalculator/Program.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Windows.Forms;
-
-namespace CircuitCalculator
-{
-	internal static class Program
-	{
-		/// <summary>
-		///     Главная точка входа для приложения.
-		/// </summary>
-		[STAThread]
-		private static void Main()
-		{
-			Application.EnableVisualStyles();
-			Application.SetCompatibleTextRenderingDefault(false);
-			Application.Run(new CalculatorForm());
-		}
-	}
-}
