+++ МЕХ/Circuit/CircuitElements/CircuitElements.csproj
@@ -42,12 +42,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Capacitor.cs" />
-    <Compile Include="Circuit.cs" />
-    <Compile Include="ElementBase.cs" />
-    <Compile Include="Inductor.cs" />
+    <Compile Include="Elements\Capacitor.cs" />
+    <Compile Include="ParentClasses\CircuitBase.cs" />
+    <Compile Include="Elements\ParallelCircuit.cs" />
+    <Compile Include="Elements\SerialCircuit.cs" />
+    <Compile Include="ParentClasses\ElementBase.cs" />
+    <Compile Include="ParentClasses\ICircuitElement.cs" />
+    <Compile Include="Elements\Inductor.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Resistor.cs" />
+    <Compile Include="Elements\Resistor.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
