+++ МЕХ/Circuit/CircuitElements/Elements/Capacitor.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Numerics;
+
+namespace CircuitElements.Elements
+{
+	/// <summary>
+	///     Конденсатор, элемент эл. цепи
+	/// </summary>
+	public class Capacitor : ElementBase
+	{
+		#region Constructor
+
+		/// <summary>
+		///     Конструктор класса Capacitor
+		/// </summary>
+		/// <param name="name"> Имя элемента </param>
+		/// <param name="value"> Значение элемента </param>
+		public Capacitor(string name, double value) : base(name, value)
+		{
+		}
+
+		#endregion
+
+		#region Public methods
+
+		/// <summary>
+		///     Расчитать импаденс по входной частоте
+		/// </summary>
+		/// <param name="frequence"> Частота сигнала </param>
+		/// <returns>Импеданс элемента в комплексной форме</returns>
+		public override Complex CalculateZ(double frequence)
+		{
+			return new Complex(0, -1 / (2 * Math.PI * frequence * Value));
+		}
+
+		#endregion
+	}
+}
