+++ МЕХ/Circuit/MainForm.sln
@@ -1,0 +1,61 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2027
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CircuitView", "MainForm\CircuitView.csproj", "{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CircuitElements", "CircuitElements\CircuitElements.csproj", "{F1FCBCA1-614D-43D3-B409-4461882EA899}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|ARM = Debug|ARM
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|ARM = Release|ARM
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|ARM.Build.0 = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|x64.Build.0 = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Debug|x86.Build.0 = Debug|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|ARM.ActiveCfg = Release|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|ARM.Build.0 = Release|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|x64.ActiveCfg = Release|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|x64.Build.0 = Release|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|x86.ActiveCfg = Release|Any CPU
+		{BAE7E8DD-7CBD-44D1-ABAB-623EBD52EFF9}.Release|x86.Build.0 = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|ARM.Build.0 = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|x64.Build.0 = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Debug|x86.Build.0 = Debug|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|ARM.ActiveCfg = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|ARM.Build.0 = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|x64.ActiveCfg = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|x64.Build.0 = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|x86.ActiveCfg = Release|Any CPU
+		{F1FCBCA1-614D-43D3-B409-4461882EA899}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F8AF051B-C8E0-4412-8E6A-50EAF269C2F5}
+	EndGlobalSection
+EndGlobal
