+++ МЕХ/Circuit/MainForm/CircuitMainForm.resx
